requirements.txt
flask
transformers
torch
openai
tiktoken
gunicorn
python-dotenv
requests
flask_cors